[Refactor] ⚛️ Allow non-integer charge values for better reconstruction
- Remove integer constraint on particle charge for higher precision
- Update function signatures to accept floating-point charge
- Improve charge calculation logic from tracker data
- Fix typo in neutron comment

+++ include/ParticleData.hh
@@ -5,7 +5,7 @@ struct ParticleData
 {
     // Particle properties
     float mass;       // Mass
-    int charge;       // Charge
+    float charge;     // Charge
     float momentum;   // Momentum
     float betaLinear; // Reconstructed beta value
